jazz report
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C53025

+++ TOneV2/Code/TOneV2/TOne.WhS.Jazz.BP.Activities/DeleteJazzReportFile.cs
@@ -9,7 +9,7 @@ namespace TOne.WhS.Jazz.BP.Activities
 {
     public sealed class DeleteJazzReportFile : CodeActivity
     {
-        public InArgument<int> FileId { get; set; }
+        public InArgument<long> FileId { get; set; }
         protected override void Execute(CodeActivityContext context)
         {
             throw new NotImplementedException();
